Moving alerts into a sub-directory.
