+++ src/alerts/enrollment-alert/hooks.js
@@ -2,8 +2,8 @@
 import {
   useContext, useState, useEffect, useCallback,
 } from 'react';
-import { UserMessagesContext, ALERT_TYPES } from '../user-messages';
-import { useModel } from '../model-store';
+import { UserMessagesContext, ALERT_TYPES } from '../../user-messages';
+import { useModel } from '../../model-store';
 import { postCourseEnrollment } from './data/api';
 
 export function useEnrollmentAlert(courseId) {
